+++ deltafi-ui/src/mocks/graphql/getAllFlows.js
@@ -38,6 +38,51 @@ const errors = [
     "message": "Action: org.deltafi.passthrough.action.RoteLoadAction is not registered as a TransformAction"
   }
 ];
+const getEnrichedData =() =>{ 
+  const data = [
+    {
+      "name": "smoke",
+      "description": "Test flow that passes data through enrich flow unchanged",
+      "sourcePlugin": {
+        "artifactId": "smoke",
+        "groupId": "org.deltafi",
+        "version": "0.17.0"
+      },
+      "flowStatus": {
+        "state": "RUNNING",
+        "errors": errors
+      },
+      "enrichActions": [
+        {
+          "name": "smoke.SmokeEnrichAction",
+          "type": "org.deltafi.passthrough.action.RoteEnrichAction",
+          "requiresDomains": [
+            "binary"
+          ],
+          "requiresEnrichment": [],
+          "requiresMetadataKeyValues": [],
+          "parameters": {
+            "enrichments": {
+              "smokeEnrichment": "smoke enrichment value"
+            }
+          }
+        }
+      ],
+      "variables": [
+        {
+          "name": "smoke-enrich-url",
+          "value": "smoke-value",
+          "description": "The URL to post the DeltaFile to",
+          "defaultValue": "http://deltafi-enrich-sink-service",
+          "dataType": "STRING"
+        }
+      ]
+    }
+  ]  
+  return data;
+}
+
+const enrichData = [[], getEnrichedData()];
 
 const generateFlows = () => {
   return {
@@ -201,6 +246,7 @@ const generateFlows = () => {
         "variables": []
       }
     ],
+    "enrich": enrichData[Math.floor(Math.random() * enrichData.length)],
     "egress": [
       {
         "name": "error",