+++ deltafi-ui/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added Pretty Print to content viewer for JSON and XML (#84)
 - Updated Trace chart to use queued vs created and added start/stop bar (#166)
 - Made content dialogs not draggable (#175)
+- Added enrich flows to the flows page (#171)
 
 ## [0.95.2]